requirements.txt
pytorch==1.10.1
torchvision==0.11.2
pillow==8.4.0
timm==0.4.12
homura-core==2021.12.1